+++ .nais/nais-dev.yaml
@@ -6,7 +6,7 @@ metadata:
   labels:
     team: poao
 spec:
-  image: docker.pkg.github.com/navikt/pto-unleash/pto-unleash:{{version}}
+  image: ghcr.io/navikt/poao-unleash/poao-unleash:{{version}}
   port: 8080
   prometheus:
     enabled: true
