+++ .nais/nais-prod.yaml
@@ -6,7 +6,7 @@ metadata:
   labels:
     team: poao
 spec:
-  image: docker.pkg.github.com/navikt/pto-unleash/pto-unleash:{{version}}
+  image: ghcr.io/navikt/poao-unleash/poao-unleash:{{version}}
   port: 8080
   prometheus:
     enabled: true
@@ -19,7 +19,7 @@ spec:
     initialDelay: 30
   replicas:
     min: 2
-    max: 4
+    max: 3
     cpuThresholdPercentage: 75
   resources:
     limits:
