+++ Dockerfile
@@ -1,2 +1,2 @@
 FROM ghcr.io/navikt/poao-baseimages/java:11
-COPY /target/pto-unleash.jar app.jar
+COPY /target/poao-unleash.jar app.jar
