+++ README.md
@@ -1,2 +1,3 @@
-# pto-unleash
-Eksponerer unleash for frontendapplikasjoner med PTO strategier
+# poao-unleash
+
+Felles tjeneste for å gjøre kall mot unleash med tilpassede strategier
