+++ src/main/java/no/nav/poao_unleash/Application.java
@@ -1,6 +1,5 @@
-package no.nav.pto_unleash;
+package no.nav.poao_unleash;
 
-import no.nav.common.utils.SslUtils;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -8,7 +7,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class Application {
 
     public static void main(String... args) {
-        SslUtils.setupTruststore();
         SpringApplication.run(Application.class, args);
     }
