+++ src/test/java/no/nav/poao_unleash/BootstrapTest.java
@@ -1,4 +1,4 @@
-package no.nav.pto_unleash;
+package no.nav.poao_unleash;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
