+++ src/test/java/no/nav/poao_unleash/TestApplicationConfig.java
@@ -1,4 +1,4 @@
-package no.nav.pto_unleash;
+package no.nav.poao_unleash;
 
 import no.nav.common.featuretoggle.UnleashClient;
 import org.springframework.context.annotation.Bean;
